% Generated by roxygen2 (4.0.1): do not edit by hand
\name{add_noise}
\alias{add_noise}
\title{Add noise / residual error to data}
\usage{
add_noise(x, ruv = list(proportional = 0, additive = 0, exponential = 0))
}
\arguments{
\item{x}{data}

\item{ruv}{list describing the magnitude of errors. List arguments: "proportional", "additive", "exponential".}
}
\description{
Add noise / residual error to data
}

