% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_sim_index_number.R
\name{add_sim_index_number}
\alias{add_sim_index_number}
\title{Add sim index number}
\usage{
add_sim_index_number(sim, id = "id")
}
\arguments{
\item{sim}{a data.frame containing the simulation data}

\item{id}{character specifying the column name in the data.frame}
}
\description{
Add simulation index number to simulation when not present
}
