% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stratification.R
\name{add_stratification}
\alias{add_stratification}
\title{Adds stratification to data set}
\usage{
add_stratification(dat, stratify, verbose = FALSE)
}
\arguments{
\item{dat}{An input data.frame or similar object}

\item{stratify}{character vector of stratification variables. Only 1 or 2 stratification variables can be supplied.}

\item{verbose}{verbosity (`TRUE` or `FALSE`)}
}
\description{
Adds stratification to data set
}
