% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_vpcdb.R
\name{as_vpcdb}
\alias{as_vpcdb}
\title{Create a vpcdb object, and standardize parameter checking}
\usage{
as_vpcdb(..., facet = NULL, scales = NULL, labeller = NULL)
}
\arguments{
\item{...}{Extra parameters (not checked) added to the object}

\item{facet}{either "wrap", "columns", or "rows"}

\item{scales}{Are scales shared across all facets (the default,
\code{"fixed"}), or do they vary across rows (\code{"free_x"}),
columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{labeller}{ggplot2 labeller function to be passed to underlying ggplot object}
}
\value{
A vpcdb object which is simply a named list with some of the values
  checked for correctness
}
\description{
Create a vpcdb object, and standardize parameter checking
}
