% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc.R
\name{calc_pred_corr_continuous}
\alias{calc_pred_corr_continuous}
\title{Perform prediction-correction}
\usage{
calc_pred_corr_continuous(
  sim,
  obs,
  pred_corr,
  pred_corr_lower_bnd,
  cols,
  verbose
)
}
\arguments{
\item{sim}{this is usually a data.frame with observed data, containing the independent and dependent variable, a column indicating the individual, and possibly covariates. E.g. load in from NONMEM using \link{read_table_nm}.  However it can also be an object like a nlmixr or xpose object}

\item{obs}{a data.frame with observed data, containing the independent and dependent variable, a column indicating the individual, and possibly covariates. E.g. load in from NONMEM using \link{read_table_nm}}

\item{pred_corr}{perform prediction-correction?}

\item{pred_corr_lower_bnd}{lower bound for the prediction-correction}

\item{verbose}{show debugging information (TRUE or FALSE)}
}
\value{
A list with "sim" and "obs" (with \code{pred_corr} performed, if
  requested)
}
\description{
Perform prediction-correction
}
