% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stratification_columns_available.R
\name{check_stratification_columns_available}
\alias{check_stratification_columns_available}
\title{Check whether stratification columns are available}
\usage{
check_stratification_columns_available(data, stratify, type = "observation")
}
\arguments{
\item{data}{`data.frame` with observation or simulation data (or \code{NULL}
to skip checking)}

\item{stratify}{vector of stratification columns (or \code{NULL} to skip
checking)}

\item{type}{either `observation` or `simulation`}
}
\value{
\code{TRUE} or raise an error about the missing columns
}
\description{
Check whether stratification columns are available
}
