% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_strat_bin.R
\name{combine_vps_strat}
\alias{combine_vps_strat}
\title{Combine a list of figures from vpc_strat_bin function}
\usage{
combine_vps_strat(plot_data)
}
\arguments{
\item{plot_data}{a list of figures and figure settings generated with 
\code{\link{vpc_strat_bin}} function}
}
\value{
a combined figure
}
\description{
\code{combine_vps_strat} combines a list of figures from 
\code{\link{vpc_strat_bin}} function
}
\details{
You can control the subplot appearance by applying ggplot functions to each
subplot items (\code{plot_data$glist[[1]]}, \code{plot_data$glist[[2]]}, ...)
}
\examples{
for (k in 1:2) plot_data$glist[[k]] + scale_y_log10()
combine_vps_strat(plot_data)


}

