% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_kmmc.R
\name{compute_kmmc}
\alias{compute_kmmc}
\title{Compute KMMC statistics}
\usage{
compute_kmmc(dat, strat = NULL, reverse_prob = FALSE, kmmc = "DOSE")
}
\arguments{
\item{dat}{data.frame with events}

\item{strat}{vector of stratification variables}

\item{reverse_prob}{reverse the probability (i.e. return `1-probability`)?}

\item{kmmc}{variable to create the KMMC plot for.}
}
\description{
Kaplan-Meier Mean Covariate plots are a simulation-based diagnostic to study the influence of covariates and identify potential model misspecification.
}
