% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_data_columns.R
\name{define_data_columns}
\alias{define_data_columns}
\title{Define data column defaults for various softwares}
\usage{
define_data_columns(sim, obs, sim_cols, obs_cols, software_type)
}
\arguments{
\item{sim}{simulated data}

\item{obs}{observed data}

\item{sim_cols}{list for mapping simulation data columns, e.g. `list(dv = "DV", id = "ID", idv = "TIME", pred="PRED")`}

\item{obs_cols}{list for mapping observation data columns, e.g. `list(dv = "DV", id = "ID", idv = "TIME", pred="PRED")`}

\item{software_type}{software type, one of `nonmem`, `phoenix`, `PKPDsim`}
}
\description{
Define data column defaults for various softwares
}
