% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_params_mvr.R
\name{draw_params_mvr}
\alias{draw_params_mvr}
\title{Draw parameters from multivariate distribution}
\usage{
draw_params_mvr(ids, n_sim, theta, omega_mat, par_names = NULL)
}
\arguments{
\item{ids}{vector of id numbers}

\item{n_sim}{number of simulations}

\item{theta}{theta vector}

\item{omega_mat}{omega matrix}

\item{par_names}{parameter names vector}
}
\description{
Draw parameters from multivariate distribution
}
