% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_dv.R
\name{filter_dv}
\alias{filter_dv}
\title{Remove values that are not observed values from data}
\usage{
filter_dv(x, verbose = FALSE, ...)
}
\arguments{
\item{verbose}{show debugging information (TRUE or FALSE)}

\item{dat}{A data.frame or similar object}
}
\value{
dat With non-observation rows removed
}
\description{
Remove values that are not observed values from data
}
