% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_vpc_input_data.R
\name{format_vpc_input_data}
\alias{format_vpc_input_data}
\title{Prepare VPC data for future calculations by standardizing column names and
modifying the input data based on the limits of quantification,
stratification, and logarithmic values.}
\usage{
format_vpc_input_data(
  dat,
  cols,
  lloq,
  uloq,
  strat,
  log_y,
  log_y_min,
  what = "observed",
  verbose = FALSE,
  pred_corr = FALSE
)
}
\arguments{
\item{dat}{An input data.frame or similar object}

\item{cols}{A list with an element for colname giving the name for colname in
\code{dat}.}

\item{lloq}{Number or NULL indicating lower limit of quantification. Default is NULL.}

\item{uloq}{Number or NULL indicating upper limit of quantification. Default is NULL.}

\item{log_y}{Boolean indicting whether y-axis should be shown as logarithmic. Default is FALSE.}

\item{log_y_min}{minimal value when using log_y argument. Default is 1e-3.}

\item{what}{The description of the data (typically "observed" or "simulated")}

\item{pred_corr}{perform prediction-correction?}
}
\value{
\code{dat} modified based on other inputs.
}
\description{
Prepare VPC data for future calculations by standardizing column names and
modifying the input data based on the limits of quantification,
stratification, and logarithmic values.
}
