% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_vpc.R
\name{ggplot.vpcdb_continuous}
\alias{ggplot.vpcdb_continuous}
\alias{ggplot.vpcdb_categorical}
\alias{ggplot.vpcdb_censored}
\alias{ggplot.vpcdb_time_to_event}
\title{Create a ggplot for each vpcdb type}
\usage{
ggplot.vpcdb_continuous(
  data = NULL,
  mapping = NULL,
  ...,
  environment = parent.frame()
)

ggplot.vpcdb_categorical(
  data = NULL,
  mapping = NULL,
  ...,
  environment = parent.frame()
)

ggplot.vpcdb_censored(
  data = NULL,
  mapping = NULL,
  ...,
  environment = parent.frame()
)

ggplot.vpcdb_time_to_event(
  data = NULL,
  mapping = NULL,
  ...,
  environment = parent.frame()
)
}
\description{
These functions are not to be called directly by users; they are for internal
use.  Users should call \code{plot_vpc()}.
}
