% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_vpc_tte.R
\name{ggplot2_label_variable}
\alias{ggplot2_label_variable}
\title{Function copied from ggplot2:::label_variable, with addition of `%||%`. 
Copied here since CRAN does not allow use of `:::`.}
\usage{
ggplot2_label_variable(labels, multi_line = TRUE)
}
\arguments{
\item{labels}{Data frame of labels. Usually contains only one element, but faceting over multiple factors entails multiple label variables.}

\item{multi_line}{Whether to display the labels of multiple factors on separate lines.}
}
\description{
Function copied from ggplot2:::label_variable, with addition of `%||%`. 
Copied here since CRAN does not allow use of `:::`.
}
