% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/guess_software.R
\name{guess_software}
\alias{guess_software}
\title{guess software package based on data}
\usage{
guess_software(software, x)
}
\arguments{
\item{software}{string specifying type of software}

\item{x}{analysis data from software}
}
\description{
guess software package based on data
}

