% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_vpc_tte.R
\name{label_vpc_tte}
\alias{label_vpc_tte}
\title{Custom ggplot2 labeller function.}
\usage{
label_vpc_tte(labels, multi_line = TRUE, sep = ": ")
}
\description{
Slight rewrite of ggplot::label_both, to make sure that labels 
for events are ordered appropriately when doing facet_wrap.
}
