% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_cens.R
\name{loq_perc}
\alias{loq_perc}
\title{Calculate percentiles below / above lloq / uloq}
\usage{
loq_perc(x, limit = 1, cens = "left")
}
\arguments{
\item{x}{data}

\item{limit}{censoring limit}

\item{cens}{censoring direction (left/right)}
}
\description{
Calculate percentiles below / above lloq / uloq
}
