% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/npde_plot.R
\name{npde_vs_time}
\alias{npde_vs_time}
\title{Create VPC-like NPDE plot (npde vs idv)}
\usage{
npde_vs_time(sim, obs, nonmem = "auto", obs_dv = "dv", obs_idv = "time",
  obs_id = "id", sim_dv = "dv", sim_idv = "time", sim_id = "id",
  n_bins = 8, smooth = TRUE, theme = "default", plot = TRUE,
  title = NULL, custom_theme = NULL)
}
\arguments{
\item{sim}{a data.frame with observed data, containing the indenpendent and dependent variable, a column indicating the individual, and possibly covariates. E.g. load in from NONMEM using \link{read_table_nm}}

\item{obs}{a data.frame with observed data, containing the indenpendent and dependent variable, a column indicating the individual, and possibly covariates. E.g. load in from NONMEM using \link{read_table_nm}}

\item{nonmem}{should variable names standard to NONMEM be used (i.e. ID, TIME, DV, PRED). Default is "auto" for autodetect}

\item{obs_dv}{variable in data.frame for observed dependent value. "dv" by default}

\item{obs_idv}{variable in data.frame for observed independent value. "time" by default}

\item{obs_id}{variable in data.frame for observed individual. "id" by default}

\item{sim_dv}{variable in data.frame for simulated dependent value. "sdv" by default}

\item{sim_idv}{variable in data.frame for simulated independent value. "time" by default}

\item{sim_id}{variable in data.frame for simulated individual. "id" by default}

\item{n_bins}{when using the "auto" binning method, what number of bins to aim for}

\item{smooth}{"smooth" the VPC (connect bin midpoints) or show bins as rectangular boxes. Default is TRUE.}

\item{theme}{which theme to load from the themes object}

\item{plot}{Boolean indacting whether to plot the ggplot2 object after creation. Default is TRUE.}

\item{title}{title}

\item{custom_theme}{specify a custom ggplot2 theme}

\item{bins}{either "auto" or a numeric vector specifying the bin separators.}

\item{auto_bin_type}{auto-binning type, default is "simple".}

\item{obs_pred}{variable in data.frame for population predicted value. "pred" by default}

\item{sim_pred}{variable in data.frame for population predicted value. "pred" by default}

\item{plot_dv}{should observations be plotted?}

\item{stratify}{character vector of stratification variables. Only 1 or 2 stratification variables can be supplied.}

\item{pi}{simulated prediction interval to plot. Default is c(0.05, 0.95),}

\item{ci}{confidence interval to plot. Default is (0.05, 0.95)}

\item{uloq}{Number or NULL indicating upper limit of quantification. Default is NULL.}

\item{lloq}{Number or NULL indicating lower limit of quantification. Default is NULL.}

\item{log_y}{Boolean indacting whether y-axis should be shown as logarithmic. Default is FALSE.}

\item{log_y_min}{minimal value when using log_y argument. Default is 1e-3.}

\item{xlab}{ylab as numeric vector of size 2}

\item{ylab}{ylab as numeric vector of size 2}

\item{facet}{either "wrap", "columns", or "rows"}
}
\value{
a list containing calculated VPC information, and a ggplot2 object
}
\description{
Creates NPDE plot from observed and simulation data
}

