% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_pred_corr_continuous.R
\name{pred_correction_core}
\alias{pred_correction_core}
\title{Core prediction correction function}
\usage{
pred_correction_core(data, pred_col, pred_corr_lower_bnd)
}
\arguments{
\item{data}{dataset, either `sim` or `obs` data.frame}

\item{pred_col}{cols$obs$pred}

\item{pred_corr_lower_bnd}{lower bound for the prediction-correction}
}
\value{
data.frame
}
\description{
Perform pred-correction for predictions that were non-zero and
were not missing observation. The latter can happen e.g. when 
censored data is set to NA in `format_vpc_input_data()`.
}
