% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replace_list_elements.R
\name{replace_list_elements}
\alias{replace_list_elements}
\title{Replace list elements by name}
\usage{
replace_list_elements(list, replacement)
}
\arguments{
\item{list}{original list}

\item{replacement}{replacement list}
}
\description{
Replace list elements by name
}
\details{
Finds and replaces list elements by name and throws an error if an 
   element is not available in the original list. This is a local duplicate
   of the PKPDmisc copy for the VPC package to reduce dependency on PKPDmisc
   at this time.
}
\examples{
\dontrun{
list <- list(ipred = "ipred", dv = "dv", idv = "idv", "pred" = "pred")
replacement <- list(dv = "conc", idv = "time")
list <- replace_list_elements(list, replacement)
}
}
