% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rtte_obs_nm}
\alias{rtte_obs_nm}
\title{Simulated RTTE data (1x)}
\format{\preformatted{'data.frame':	573 obs. of  6 variables:
 $ id  : num  1 1 2 2 2 2 2 3 4 4 ...
 $ t   : num  21 100 3 52 68 74 100 100 8 35 ...
 $ dv  : num  1 0 1 1 1 1 0 0 1 1 ...
 $ dt  : num  21 79 3 49 16 6 26 100 8 26 ...
 $ drug: num  1 1 1 1 1 1 1 1 1 1 ...
 $ sex : num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
rtte_obs_nm
}
\description{
An example dataset with simulated repeated time-to-event data
}
\keyword{datasets}

