% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{rtte_sim_nm}
\alias{rtte_sim_nm}
\title{Simulated RTTE data (100x)}
\format{\preformatted{'data.frame':	2000000 obs. of  7 variables:
 $ sim : int  1 1 1 1 1 1 1 1 1 1 ...
 $ id  : num  1 1 1 1 1 1 1 1 1 1 ...
 $ t   : num  1 2 3 4 5 6 7 8 9 10 ...
 $ dv  : num  0 0 0 0 0 0 0 0 0 0 ...
 $ rtte: num  0 0 0 0 0 0 0 0 0 0 ...
 $ drug: num  1 1 1 1 1 1 1 1 1 1 ...
 $ sex : num  1 1 1 1 1 1 1 1 1 1 ...
}}
\usage{
rtte_sim_nm
}
\description{
An example dataset with simulated repeated time-to-event data (100 simulations)
}
\keyword{datasets}

