% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_default.R
\docType{data}
\name{show_default}
\alias{show_default}
\title{Defaults for show argument with various endpoint types}
\format{
An object of class \code{list} of length 4.
}
\usage{
show_default
}
\description{
Defaults for show argument with various endpoint types
}
\keyword{datasets}
