% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_default.R
\docType{data}
\name{show_default_tte}
\alias{show_default_tte}
\title{Defaults for show argument for TTE VPC}
\format{An object of class \code{list} of length 11.}
\usage{
show_default_tte
}
\description{
Defaults for show argument for TTE VPC
}
\keyword{datasets}
