% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_data}
\alias{sim_data}
\title{Simulate data based on a model and parameter distributions}
\usage{
sim_data(design = cbind(id = c(1, 1, 1), idv = c(0, 1, 2)), model, theta,
  omega_mat, par_names, n = 100)
}
\arguments{
\item{design}{}

\item{model}{}

\item{theta}{}

\item{omega_mat}{}

\item{par_names}{A vector describing}
}
\value{
A vector of simulated dependent variables (for us in the VPC plotting function)
}
\description{
Simulate data based on a model and parameter distributions
}
\details{
This function generates the simulated dependent values for use in the VPC plotting function.
}
\examples{
obs <- Theoph
colnames(obs) <- c("id", "wt", "dose", "time", "dv")
obs$sex <- round(runif(unique(obs$id))) # create a dummy covariate to show stratification

sim <- sim_data(obs, # the design of the dataset
                model = function(x, par) { # the model
                  pk_oral_1cmt (t = x$time, dose=x$dose * x$wt, ka = par$ka, ke = par$ke, cl = par$cl * x$wt, ruv = list(additive = 0.1))
                },
                theta = c(2.774, 0.0718, .0361),                 # parameter values
                omega_mat = c(0.08854,                           # specified as lower triangle by default;
                              0.02421, 0.02241,                  # note: assumed that every theta has iiv, set to 0 if no iiv.
                              0.008069, 0.008639, 0.02862),
                par_names = c("ka", "ke", "cl"),                 # link the parameters in the model to the thetas/omegas
                n = 500)

vpc_dat <- plot_vpc(sim, obs,
                    bins = c(0, 2, 4, 6, 8, 10, 25),
                    strat = "sex",
                    ylab = "Concentration", xlab = "Time (hrs)", title="Visual predictive check")
}
\seealso{
\code{\link{plot_vpc}}
}

