% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{simple_data}
\alias{simple_data}
\title{A small rich dataset}
\format{An object of class \code{list} of length 2.}
\usage{
simple_data
}
\description{
A small rich dataset
}
\details{
a list containing the obs and sim data for an example dataset to run a 
simple vpc.
}
\examples{
\dontrun{
vpc(simple_data$sim, simple_data$obs)
}
}
\keyword{datasets}
