% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\name{theme_plain}
\alias{theme_plain}
\title{A nicer default theme for ggplot2}
\usage{
theme_plain()
}
\description{
A nicer default theme for ggplot2
}
\examples{
vpc(simple_data$sim, simple_data$obs) + theme_plain()

}
