% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/triangle_to_full.R
\name{triangle_to_full}
\alias{triangle_to_full}
\title{Lower to full triangle}
\usage{
triangle_to_full(vect)
}
\arguments{
\item{vect}{the lower triangle of a covariance matrix}
}
\description{
Convert the lower triangle of a covariance matrix to a full matrix object
}
