% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{vpc-package}
\alias{vpc-package}
\title{VPC package}
\description{
Create Visual Predictive Checks from within R, either from a specified model and parameter distributions, or from simulated data.
}
\author{
Ron Keizer \email{ronkeizer@gmail.com}
}

