% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vpc_loq}
\alias{vpc_loq}
\title{VPC function for LOQ-type data}
\usage{
vpc_loq(sim, obs, type = "lloq", bins, strat = NULL, dv = "dv",
  idv = "time", pi = c(0.05, 0.95), ci = c(0.05, 0.95), uloq = NULL,
  lloq = NULL)
}
\arguments{
\item{sim}{}

\item{obs}{}
}
\value{
Either the data for plotting a VPC or a ggplot2 object
}
\description{
Creates a VPC plot and/or plotting data from observed and simulation data
}

