% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_strat_bin.R
\name{vpc_strat_bin}
\alias{vpc_strat_bin}
\title{Create VPC plot with different binning settings for each stratification group}
\usage{
vpc_strat_bin(sim = NULL, obs = NULL, bins = "jenks", bin_list = NULL,
  n_bins = "auto", n_bins_list = NULL, stratify = NULL,
  strat_fig_dir = 1, facet_names = TRUE, verbose = FALSE, xylab = FALSE,
  return_fig = TRUE, ...)
}
\arguments{
\item{sim}{see \code{\link{vpc}} function}

\item{obs}{see \code{\link{vpc}} function}

\item{bins}{see \code{\link{vpc}} function}

\item{bin_list}{a list to assign \code{bins} for each stratification group}

\item{n_bins}{see \code{\link{vpc}} function}

\item{n_bins_list}{a vector to assign \code{n_bins} for each stratification group}

\item{stratify}{see \code{\link{vpc}} function. The name of the column cannot be 
"strat" or "strat_for_title"}

\item{strat_fig_dir}{set the direction of figure arrangement for two 
stratification variables. If strat_fig_dir == 1 (default), first variable 
will be spread for horizontal direction}

\item{facet_names}{see \code{\link{vpc}} function}

\item{verbose}{show debugging information (TRUE or FALSE)}

\item{xylab}{boolean indicting whether to show axis labels for individual subplots}

\item{return_fig}{boolean whether to return a combined figure (TRUE, default) or a list object of individual plots (FALSE)}

\item{...}{other inputs passed on to \code{\link{vpc}} function}
}
\value{
a combined figure (return_fig==TRUE, default) or a list object of individual plots (return_fig==FALSE)
}
\description{
\code{vpc_strat_bin} provides a workaround to create a VPC plot with
different bin settings for each stratification group of data.
}
\details{
To control the graph settings with ggplot2 syntax, you can set 
\code{return_fig = FALSE}, edit each subplot in \code{plot_data$glist} with ggplot2,
and apply \code{\link{combine_vps_strat}} function to create combined plot.
}
\examples{
vpc_strat_bin(sim = simple_data$sim,
              obs = simple_data$obs,
              stratify = "ISM", 
              n_bins_list = c(4,8))
              
plot_data <- 
  vpc_strat_bin(sim = simple_data$sim,
                obs = simple_data$obs,
                stratify = "ISM", 
                n_bins_list = c(4,8),
                return_fig = F)


}

