% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpc_theme.R
\name{vpc_theme}
\alias{vpc_theme}
\title{VPC theming}
\usage{
vpc_theme(update = NULL)
}
\arguments{
\item{update}{list, possible arguments: obs_color, obs_size, obs_median_color, obs_median_linetype, obs_mdian_size, obs_ci_color, obs_ci_linetype, obs_ci_size, sim_pi_fill, sim_pi_alpha, sim_pi_color, sim_pi_linetype, sim_pi_size, sim_median_fill, sim_median_alpha, sim_median_color, sim_median_linetype, sim_median_size, bin_separators_color}
}
\description{
VPC theming
}

