% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vpc_tte}
\alias{vpc_tte}
\title{VPC function for survival-type data}
\usage{
vpc_tte(sim, obs, rtte = FALSE, rtte_flag = "rtte", occasions = NULL,
  bin_method = "obs", filter_rtte = "rtte", n_bins = 32, n_sim = "auto",
  sim_dense_grid = FALSE, auto_bin_type = "simple", obs_dv = "dv",
  sim_dv = "sdv", obs_idv = "time", sim_idv = "time", obs_id = "id",
  sim_id = "id", pi_med = FALSE, nonmem = FALSE, stratify = NULL,
  ci = c(0.05, 0.95), plot = TRUE, xlab = NULL, ylab = NULL,
  title = NULL, smooth = FALSE, theme = "default", custom_theme = NULL,
  facet = "wrap")
}
\arguments{
\item{sim}{}

\item{obs}{}
}
\value{
A VPC object
}
\description{
Creates a VPC plot and/or plotting data from observed and simulation data
}
\examples{
## Example for repeated) time-to-event data
## with NONMEM-like data (e.g. simulated using a dense grid)
data(rtte_obs_nm)
data(rtte_sim_nm)

# treat RTTE as TTE, no stratification
vpc_tte(rtte_sim_nm, rtte_obs_nm,
        rtte = FALSE, bin_method="obs",
        sim_dv = "dv", obs_idv = "t", sim_idv = "t", n_sim = 100)

# stratified for covariate and study arm
vpc_tte(rtte_sim_nm, rtte_obs_nm,
        stratify = c("sex","drug"),
        rtte = FALSE, bin_method = "spread", n_bins=16,
        sim_dv = "dv", obs_idv = "t", sim_idv = "t", n_sim = 100)

# stratified per event number (we'll only look at first 3 events) and stratify per arm
vpc_tte(rtte_sim_nm, rtte_obs_nm,
        rtte = TRUE, occasions = c(1:3),
        stratify = c("drug"), bin_method="obs",
        sim_dv = "dv", obs_idv = "t", sim_idv = "t", n_sim = 100)
}

