% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_noise.R
\name{add_noise}
\alias{add_noise}
\title{Add noise / residual error to data}
\usage{
add_noise(x, ruv = list(proportional = 0, additive = 0, exponential = 0))
}
\arguments{
\item{x}{data}

\item{ruv}{list describing the magnitude of errors. List arguments: "proportional", "additive", "exponential".}
}
\description{
Add noise / residual error to data
}
\examples{
library(dplyr)
ipred <- c(10, 8, 6, 4, 2, 0) \%>\% add_noise(ruv = list(proportional = 0.1, additive = 0.2))
}
