% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_stratification.R
\name{add_stratification}
\alias{add_stratification}
\title{Adds stratification to data set}
\usage{
add_stratification(dat, strat, verbose = FALSE)
}
\arguments{
\item{dat}{data.frame}

\item{strat}{vector of stratification variables}

\item{verbose}{verbosity (`TRUE` or `FALSE`)}
}
\description{
Adds stratification to data set
}
