% Generated by roxygen2 (4.0.1): do not edit by hand
\name{auto_bin}
\alias{auto_bin}
\title{Calculate appropriate bin separators for vpc}
\usage{
auto_bin(dat, type = "simple", n_bins = 8, x = "time")
}
\arguments{
\item{dat}{data}

\item{type}{auto-binning type. Currently only "simple" available.}

\item{n_bins}{number of bins to aim for.}

\item{x}{name of independent variable in dataset}
}
\value{
A vector of bin separators
}
\description{
Calculate appropriate bin separators for vpc
}
\details{
This function calculates bin separators (e.g. for use in a vpc) based on nadirs in the density functions for the indenpendent variable
}
\seealso{
\code{\link{vpc}}
}

