% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/binning.R
\name{auto_bin}
\alias{auto_bin}
\title{Calculate appropriate bin separators for vpc}
\usage{
auto_bin(dat, type = "kmeans", n_bins = 8, verbose = FALSE)
}
\arguments{
\item{dat}{data frame}

\item{type}{auto-binning type: "density", "time", or "data"}

\item{n_bins}{number of bins to use. For "density" the function might not return a solution with the exact number of bins.}

\item{verbose}{show warnings and other messages (TRUE or FALSE)}
}
\value{
A vector of bin separators
}
\description{
This function calculates bin separators (e.g. for use in a vpc) based on nadirs in the density functions for the independent variable
}
