% Generated by roxygen2 (4.0.1): do not edit by hand
\name{bin_data}
\alias{bin_data}
\title{Function to bin data based on a vector of bin separators, e.g. for use in VPC}
\usage{
bin_data(x, bins = c(0, 3, 5, 7), idv = "time")
}
\arguments{
\item{x}{data}

\item{bins}{numeric vector specifying bin separators}

\item{idv}{variable in the data specifies the independent variable (e.g. "time")}
}
\description{
Function to bin data based on a vector of bin separators, e.g. for use in VPC
}

