% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_stratification_columns_available.R
\name{check_stratification_columns_available}
\alias{check_stratification_columns_available}
\title{Check whether stratification columns are available}
\usage{
check_stratification_columns_available(data, stratify, type = "observation")
}
\arguments{
\item{data}{`data.frame` with observation or simulation data}

\item{stratify}{vector of stratification columns}

\item{type}{either `observation` or `simulation`}
}
\description{
Check whether stratification columns are available
}
