% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_kaplan.R
\name{compute_kaplan}
\alias{compute_kaplan}
\title{Compute Kaplan-Meier statistics}
\usage{
compute_kaplan(dat, strat = "strat", reverse_prob = FALSE, ci = NULL)
}
\arguments{
\item{dat}{data.frame with events}

\item{strat}{vector of stratification variables}

\item{reverse_prob}{reverse the probability (i.e. return `1-probability`)?}

\item{ci}{confidence interval to calculate, numeric vector of length 2}
}
\description{
Compute Kaplan-Meier statistics
}
