% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vpc_theme.R
\name{create_vpc_theme}
\alias{create_vpc_theme}
\title{Create new vpc theme}
\usage{
create_vpc_theme(...)
}
\arguments{
\item{...}{pass arguments to `new_vpc_theme`}
}
\description{
Create new vpc theme
}
