% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_cens.R
\name{loq_frac}
\alias{loq_frac}
\title{Calculate fraction of observations below lloq / above uloq}
\usage{
loq_frac(x, limit = 1, cens = c("left", "right", "neither", "both"))
}
\arguments{
\item{x}{A numeric vector}

\item{limit}{censoring limit (ignored if \code{cens="neither"})}

\item{cens}{censoring direction}
}
\value{
The fraction of observations (\code{NA} is counted as below/above)
}
\description{
Calculate fraction of observations below lloq / above uloq
}
