% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile_cens.R
\name{quantile_cens}
\alias{quantile_cens}
\title{Calculate quantiles respecting the censored data}
\usage{
quantile_cens(x, p = 0.5, limit = 1, cens = "left")
}
\arguments{
\item{x}{data}

\item{p}{quantile}

\item{limit}{censoring limit}

\item{cens}{censoring direction (left/right)}
}
\description{
Calculate quantiles respecting the censored data
}
