% Generated by roxygen2 (4.0.2): do not edit by hand
\name{read_table_nm}
\alias{read_table_nm}
\title{Read table file simulated with NONMEM}
\usage{
read_table_nm(file, perl = TRUE)
}
\arguments{
\item{file}{The file to be read}

\item{perl}{a boolean parameter indicating whether to use Perl to read in the file. Default is TRUE. Use of Perl is advised as it is much faster.}
}
\value{
a data.frame
}
\description{
\code{read_table_nm} will read any table simulated from NONMEM in the $TABLE record. Tables can be simulated using either ONEHEADER or NOHEADER, and can be single or multiple subproblems.
}
\seealso{
\link{vpc}
}

