% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/read_table_nm.R
\name{read_table_nm}
\alias{read_table_nm}
\title{Read table file simulated with NONMEM}
\usage{
read_table_nm(file, perl = NULL)
}
\arguments{
\item{file}{The file to be read}

\item{perl}{a boolean or NULL variable indicating whether to use Perl to read in the file. Default is NULL, which will make the function check whether Perl is available. Use of Perl is advised as it is much faster.}
}
\value{
a data.frame
}
\description{
\code{read_table_nm} will read any table simulated from NONMEM in the $TABLE record. Tables can be simulated using either ONEHEADER or NOHEADER, and can be single or multiple subproblems.
}
\seealso{
\link{vpc}
}

