% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rtte_obs_nm}
\alias{rtte_obs_nm}
\title{Simulated RTTE data (1x)}
\format{An object of class \code{data.frame} with 573 rows and 6 columns.}
\usage{
rtte_obs_nm
}
\description{
An example dataset with simulated repeated time-to-event data
}
\keyword{datasets}
