% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{rtte_sim_nm}
\alias{rtte_sim_nm}
\title{Simulated RTTE data (100x)}
\format{
An object of class \code{data.frame} with 2000000 rows and 7 columns.
}
\usage{
rtte_sim_nm
}
\description{
An example dataset with simulated repeated time-to-event data (100 simulations)
}
\keyword{datasets}
