% Generated by roxygen2 (4.0.1): do not edit by hand
\name{sim_data}
\alias{sim_data}
\title{Simulate data based on a model and parameter distributions}
\usage{
sim_data(design = cbind(id = c(1, 1, 1), idv = c(0, 1, 2)),
  model = function(x) {     return(x$alpha + x$beta) }, theta, omega_mat,
  par_names, par_values = NULL, draw_iiv = "mvrnorm",
  error = list(proportional = 0, additive = 0, exponential = 0), n = 100)
}
\arguments{
\item{design}{}

\item{model}{A function with the first argument the simulation design, i.e. a dataset with the columns ... The second argument to this function is a dataset with parameters for every individual. This can be supplied by the user, or generated by this sim_data if theta and omega_mat are supplied.}

\item{theta}{}

\item{omega_mat}{}

\item{par_names}{A vector describing}
}
\value{
A vector of simulated dependent variables (for us in the VPC plotting function)
}
\description{
Simulate data based on a model and parameter distributions
}
\details{
This function generates the simulated dependent values for use in the VPC plotting function.
}
\examples{
obs <- Theoph
colnames(obs) <- c("id", "wt", "dose", "time", "dv")
obs <- obs \%>\%   # create a dummy covariate to show stratification
 group_by(id) \%>\%
 mutate(sex = round(runif(1)))

sim <- sim_data(obs, # the design of the dataset
                model = function(x) { # the model
                  pk_oral_1cmt (t = x$time, dose=x$dose * x$wt, ka = x$ka, ke = x$ke, cl = x$cl * x$wt, ruv = list(additive = 0.1))
                },
                theta = c(2.774, 0.0718, .0361),                 # parameter values
                omega_mat = c(0.08854,                           # specified as lower triangle by default;
                              0.02421, 0.02241,                  # note: assumed that every theta has iiv, set to 0 if no iiv.
                              0.008069, 0.008639, 0.02862),
                par_names = c("ka", "ke", "cl"),                 # link the parameters in the model to the thetas/omegas
                n = 500)

vpc_dat <- vpc(sim, obs, stratify = c("sex"),
              plot.dv = FALSE, facet = "wrap",
              ylab = "Concentration", xlab = "Time (hrs)")
}
\seealso{
\code{\link{vpc}}
}

