% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vpc}
\alias{vpc}
\title{VPC function}
\usage{
vpc(sim = NULL, obs = NULL, bins = "jenks", n_bins = "auto",
  obs_dv = NULL, sim_dv = NULL, obs_idv = NULL, sim_idv = NULL,
  obs_id = NULL, sim_id = NULL, obs_pred = NULL, sim_pred = NULL,
  nonmem = "auto", plot = FALSE, plot_dv = FALSE, plot_obs_ci = TRUE,
  plot_pi_ci = TRUE, plot_obs_median = TRUE, plot_sim_median = TRUE,
  plot_sim_median_ci = TRUE, stratify = NULL, stratify_color = NULL,
  legend_pos = NULL, pred_corr = FALSE, pred_corr_lower_bnd = 0,
  pi = c(0.05, 0.95), ci = c(0.05, 0.95), uloq = NULL, lloq = NULL,
  log_y = FALSE, log_y_min = 0.001, xlab = NULL, ylab = NULL,
  title = NULL, smooth = TRUE, theme = "default", custom_theme = NULL,
  facet = "wrap")
}
\arguments{
\item{sim}{a data.frame with observed data, containing the indenpendent and dependent variable, a column indicating the individual, and possibly covariates. E.g. load in from NONMEM using \link{read_table_nm}}

\item{obs}{a data.frame with observed data, containing the indenpendent and dependent variable, a column indicating the individual, and possibly covariates. E.g. load in from NONMEM using \link{read_table_nm}}

\item{bins}{either "auto" or a numeric vector specifying the bin separators.}

\item{n_bins}{when using the "auto" binning method, what number of bins to aim for}

\item{obs_dv}{variable in data.frame for observed dependent value. "dv" by default}

\item{sim_dv}{variable in data.frame for simulated dependent value. "sdv" by default}

\item{obs_idv}{variable in data.frame for observed independent value. "time" by default}

\item{sim_idv}{variable in data.frame for simulated independent value. "time" by default}

\item{obs_id}{variable in data.frame for observed individual. "id" by default}

\item{sim_id}{variable in data.frame for simulated individual. "id" by default}

\item{obs_pred}{variable in data.frame for population predicted value. "pred" by default}

\item{sim_pred}{variable in data.frame for population predicted value. "pred" by default}

\item{nonmem}{should variable names standard to NONMEM be used (i.e. ID, TIME, DV, PRED). Default is "auto" for autodetect}

\item{plot}{Boolean indacting whether to plot the ggplot2 object after creation. Default is FALSE.}

\item{plot_dv}{should observations be plotted?}

\item{plot_obs_ci}{default is TRUE}

\item{plot_pi_ci}{default is TRUE}

\item{plot_obs_median}{default is TRUE}

\item{plot_sim_median}{default is TRUE}

\item{plot_sim_median_ci}{default is TRUE}

\item{stratify}{character vector of stratification variables. Only 1 or 2 stratification variables can be supplied.}

\item{stratify_color}{variable to stratify and color lines for observed data. Only 1 stratification variables can be supplied.}

\item{pred_corr}{perform prediction-correction?}

\item{pred_corr_lower_bnd}{lower bound for the prediction-correction}

\item{pi}{simulated prediction interval to plot. Default is c(0.05, 0.95),}

\item{ci}{confidence interval to plot. Default is (0.05, 0.95)}

\item{uloq}{Number or NULL indicating upper limit of quantification. Default is NULL.}

\item{lloq}{Number or NULL indicating lower limit of quantification. Default is NULL.}

\item{log_y}{Boolean indacting whether y-axis should be shown as logarithmic. Default is FALSE.}

\item{log_y_min}{minimal value when using log_y argument. Default is 1e-3.}

\item{xlab}{ylab as numeric vector of size 2}

\item{ylab}{ylab as numeric vector of size 2}

\item{title}{title}

\item{smooth}{"smooth" the VPC (connect bin midpoints) or show bins as rectangular boxes. Default is TRUE.}

\item{theme}{which theme to load from the themes object}

\item{custom_theme}{specify a custom ggplot2 theme}

\item{facet}{either "wrap", "columns", or "rows"}

\item{bins}{either "density", "time", or "data", or a numeric vector specifying the bin separators.}
}
\value{
a list containing calculated VPC information, and a ggplot2 object
}
\description{
Creates a VPC plot from observed and simulation data
}
\examples{
obs <- Theoph
colnames(obs) <- c("id", "wt", "dose", "time", "dv")
obs <- obs \%>\%   # create a dummy covariate to show stratification
 group_by(id) \%>\%
 mutate(sex = round(runif(1)))

sim <- sim_data(obs, # the design of the dataset
                model = function(x) { # the model
                  pk_oral_1cmt (t = x$time, dose=x$dose * x$wt, ka = x$ka,
                                ke = x$ke, cl = x$cl * x$wt,
                                ruv = list(additive = 0.1))
                },
                theta = c(2.774, 0.0718, .0361),             # parameter values
                omega_mat = c(0.08854,                       # specified as lower triangle
                              0.02421, 0.02241,              # note: assumed every th has iiv,
                              0.008069, 0.008639, 0.02862),  #  set to 0 if no iiv.
                par_names = c("ka", "ke", "cl"),             # link the parameters in the model
                n = 500)                                     #   to the thetas/omegas

vpc_dat <- vpc(sim, obs, stratify = c("sex"))
}
\seealso{
\link{sim_data}, \link{vpc_cens}, \link{vpc_tte}
}

