% Generated by roxygen2 (4.0.2): do not edit by hand
\name{vpc_cens}
\alias{vpc_cens}
\title{VPC function for left- or right-censored data (e.g. BLOQ data)}
\usage{
vpc_cens(sim = NULL, obs = NULL, bins = "jenks", n_bins = 8,
  obs_dv = NULL, sim_dv = NULL, obs_idv = NULL, sim_idv = NULL,
  obs_id = NULL, sim_id = NULL, nonmem = "auto", stratify = NULL,
  stratify_color = NULL, ci = c(0.05, 0.95), uloq = NULL, lloq = NULL,
  plot = FALSE, xlab = NULL, ylab = NULL, plot_sim_med = FALSE,
  title = NULL, smooth = TRUE, theme = "default", custom_theme = NULL,
  facet = "wrap")
}
\arguments{
\item{sim}{a data.frame with observed data, containing the indenpendent and dependent variable, a column indicating the individual, and possibly covariates. E.g. load in from NONMEM using \link{read_table_nm}}

\item{obs}{a data.frame with observed data, containing the indenpendent and dependent variable, a column indicating the individual, and possibly covariates. E.g. load in from NONMEM using \link{read_table_nm}}

\item{bins}{either "density", "time", or "data", or a numeric vector specifying the bin separators.}

\item{n_bins}{number of bins}

\item{obs_dv}{variable in data.frame for observed dependent value. "dv" by default}

\item{sim_dv}{variable in data.frame for simulated dependent value. "sdv" by default}

\item{obs_idv}{variable in data.frame for observed independent value. "time" by default}

\item{sim_idv}{variable in data.frame for simulated independent value. "time" by default}

\item{obs_id}{variable in data.frame for observed individual. "id" by default}

\item{sim_id}{variable in data.frame for simulated individual. "id" by default}

\item{nonmem}{should variable names standard to NONMEM be used (i.e. ID, TIME, DV, PRED).  Default is "auto" for autodetect}

\item{stratify}{character vector of stratification variables. Only 1 or 2 stratification variables can be supplied.}

\item{stratify_color}{variable to stratify and color lines for observed data. Only 1 stratification variables can be supplied.}

\item{ci}{confidence interval to plot. Default is (0.05, 0.95)}

\item{uloq}{Number or NULL indicating upper limit of quantification. Default is NULL.}

\item{lloq}{Number or NULL indicating lower limit of quantification. Default is NULL.}

\item{plot}{Boolean indacting whether to plot the ggplot2 object after creation. Default is FALSE.}

\item{xlab}{ylab as numeric vector of size 2}

\item{ylab}{ylab as numeric vector of size 2}

\item{title}{title}

\item{smooth}{"smooth" the VPC (connect bin midpoints) or show bins as rectangular boxes. Default is TRUE.}

\item{theme}{which theme to load from the themes object}

\item{custom_theme}{specify a custom ggplot2 theme}

\item{facet}{either "wrap", "columns", or "rows"}

\item{type}{either "lloq" (default) or "uloq".}
}
\value{
a list containing calculated VPC information, and a ggplot2 object
}
\description{
Creates a VPC plot from observed and simulation data
sim,
}
\examples{
obs <- Theoph
colnames(obs) <- c("id", "wt", "dose", "time", "dv")
obs <- obs \%>\%   # create a dummy covariate to show stratification
 group_by(id) \%>\%
 mutate(sex = round(runif(1)))

sim <- sim_data(obs, # the design of the dataset
                model = function(x) { # the model
                  pk_oral_1cmt (t = x$time, dose=x$dose * x$wt, ka = x$ka,
                                ke = x$ke, cl = x$cl * x$wt,
                                ruv = list(additive = 0.1))
                },
                theta = c(2.774, 0.0718, .0361),             # parameter values
                omega_mat = c(0.08854,                       # specified as lower triangle
                              0.02421, 0.02241,              # note: assumed every th has iiv,
                              0.008069, 0.008639, 0.02862),  #  set to 0 if no iiv.
                par_names = c("ka", "ke", "cl"),             # link the parameters in the model
                n = 500)                                     #   to the thetas/omegas

vpc_loq <- vpc_cens(sim, obs, lloq = 5)
}
\seealso{
\link{vpc}
}

