% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vpc_loq}
\alias{vpc_loq}
\title{VPC function for left- or right-censored data (e.g. BLOQ data)}
\usage{
vpc_loq(sim, obs, bins = NULL, n_bins = 8, obs.dv = "dv",
  sim.dv = "sdv", obs.idv = "time", sim.idv = "time", stratify = NULL,
  pi = c(0.05, 0.95), ci = c(0.05, 0.95), uloq = NULL, lloq = NULL,
  plot = TRUE, log_y = FALSE, xlab = NULL, ylab = NULL, title = NULL,
  smooth = TRUE, theme = "default", custom_theme = NULL,
  return_what = "data", type = "bloq")
}
\arguments{
\item{sim}{}

\item{obs}{}
}
\value{
Either the data for plotting a VPC or a ggplot2 object
}
\description{
Creates a VPC plot and/or plotting data from observed and simulation data
}

