% Generated by roxygen2 (4.0.1): do not edit by hand
\name{vpc_tte}
\alias{vpc_tte}
\title{VPC function for survival-type data}
\usage{
vpc_tte(sim = NULL, obs = NULL, psn_folder = NULL, rtte = FALSE,
  rtte_calc_diff = TRUE, events = NULL, bins = FALSE, n_bins = 10,
  software = "auto", obs_cols = NULL, sim_cols = NULL, kmmc = NULL,
  reverse_prob = FALSE, stratify = NULL, stratify_color = NULL,
  ci = c(0.05, 0.95), plot = FALSE, xlab = NULL, ylab = NULL,
  show = list(), as_percentage = TRUE, title = NULL, smooth = FALSE,
  vpc_theme = NULL, ggplot_theme = NULL, facet = "wrap",
  verbose = FALSE)
}
\arguments{
\item{sim}{a data.frame with observed data, containing the indenpendent and dependent variable, a column indicating the individual, and possibly covariates. E.g. load in from NONMEM using \link{read_table_nm}}

\item{obs}{a data.frame with observed data, containing the indenpendent and dependent variable, a column indicating the individual, and possibly covariates. E.g. load in from NONMEM using \link{read_table_nm}}

\item{rtte}{repeated time-to-event data? Deafult is FALSE (treat as single-event TTE)}

\item{rtte_calc_diff}{recalculate time (T/F)? When simulating in NONMEM, you will probably need to set this to TRUE to recalculate the TIME to relative times between events (unless you output the time difference between events and specify that as independent variable to the vpc_tte() funciton.}

\item{events}{numeric vector describing which events to show a VPC for when repeated TTE data, e.g. c(1:4). Default is NULL, which shows all events.}

\item{nonmem}{should variable names standard to NONMEM be used (i.e. ID, TIME, DV, PRED). Default is "auto" for autodetect.}

\item{dense_grid}{Was a dense grid used to simulate and should the vpc_tte function remove the non-event data? TRUE or FALSE. Either way should not affect plot, when points removed the plot will just be generated faster.}

\item{reverse_prob}{reverse the probability scale (i.e. plot 1-probability)}

\item{stratify}{character vector of stratification variables. Only 1 or 2 stratification variables can be supplied. If stratify_color is also specified, only 1 additional stratification can be specified.}

\item{stratify_color}{variable to stratify and color lines for observed data. Only 1 stratification variables can be supplied.}

\item{ci}{confidence interval to plot. Default is (0.05, 0.95)}

\item{plot}{Boolean indacting whether to plot the ggplot2 object after creation. Default is FALSE.}

\item{xlab}{ylab as numeric vector of size 2}

\item{ylab}{ylab as numeric vector of size 2}

\item{title}{title}

\item{smooth}{"smooth" the VPC (connect bin midpoints) or show bins as rectangular boxes. Default is TRUE.}

\item{vpc_theme}{theme to be used in VPC. Expects list of class vpc_theme created with function vpc_theme()}

\item{ggplot_theme}{specify a custom ggplot2 theme}

\item{facet}{either "wrap", "columns", or "rows"}

\item{verbose}{TRUE or FALSE (default)}
}
\value{
a list containing calculated VPC information, and a ggplot2 object
}
\description{
Creates a VPC plot from observed and simulation survival data
}
\examples{
## Example for repeated) time-to-event data
## with NONMEM-like data (e.g. simulated using a dense grid)
data(rtte_obs_nm)
data(rtte_sim_nm)

# treat RTTE as TTE, no stratification
vpc_tte(sim = rtte_sim_nm,
        obs = rtte_obs_nm,
        rtte = FALSE,
        sim_cols=list(dv = "dv", idv = "t"), obs_cols=list(idv = "t"))

# stratified for covariate and study arm
vpc_tte(sim = rtte_sim_nm,
        obs = rtte_obs_nm,
        stratify = c("sex","drug"),
        rtte = FALSE,
        sim_cols=list(dv = "dv", idv = "t"), obs_cols=list(idv = "t"))

# stratified per event number (we'll only look at first 3 events) and stratify per arm
vpc_tte(sim = rtte_sim_nm,
        obs = rtte_obs_nm,
        rtte = TRUE, events = c(1:3),
        stratify = c("drug"),
        sim_cols=list(dv = "dv", idv = "t"), obs_cols=list(idv = "t"))
}
\seealso{
\link{vpc}
}

